refactor: fees component created
