+++ src/components/games/FilteredGameCard.tsx
@@ -2,7 +2,7 @@ import time from "../../assets/v11/time.svg";
 import participants from "../../assets/v11/participants.svg";
 import { Game } from "../../interfaces/games.ts";
 import { Link } from "react-router-dom";
-import { GameStatus } from "../../features/games/GameStatus.tsx";
+import { GameStatus } from "../../features/games/components/GameStatus.tsx";
 
 interface FilteredGameCardProp {
   game: Game;
