+++ src/components/games/GameListContainer.tsx
@@ -16,11 +16,9 @@ const GameListContainer = ({
     <div className="sm:hidden md:block custom-scrollbar bg-light-dark sm:w-full md:w-[381px] sm:h-[33.25rem] md:h-[494px] sm:p-[0.5rem] md:p-4 rounded-[20px] space-y-3 overflow-y-scroll">
       {isLoading && (
         <>
-          <GameCardSkeleton />
-          <GameCardSkeleton />
-          <GameCardSkeleton />
-          <GameCardSkeleton />
-          <GameCardSkeleton />
+          {Array.from({ length: 5 }).map((_, index) => (
+            <GameCardSkeleton key={index} />
+          ))}
         </>
       )}
       {allGamesData?.length === 0 ? (
