+++ src/features/games/components/GameCardLink.tsx
@@ -1,0 +1,3 @@
+export const GameCardLink = () => {
+  return <div>GameCardLink</div>;
+};
